Put normal and sparse merkle into one folder and extract common functions
