+++ eth2/_utils/blobs.py
@@ -16,7 +16,7 @@ from eth_utils import (
 from eth._utils.padding import (
     zpad_right,
 )
-from eth2._utils.merkle_normal import (
+from eth2._utils.merkle.normal import (
     get_merkle_root_from_items,
 )
