+++ eth2/_utils/merkle/normal.py
@@ -8,7 +8,6 @@ not considered to be part of the tree.
 import math
 from typing import (
     Iterable,
-    NewType,
     Sequence,
     Union,
 )
@@ -16,7 +15,6 @@ from typing import (
 from cytoolz import (
     identity,
     iterate,
-    partition,
     reduce,
     take,
 )
@@ -30,23 +28,14 @@ from eth_utils import (
     ValidationError,
 )
 
-
-MerkleTree = NewType("MerkleTree", Sequence[Sequence[Hash32]])
-MerkleProof = NewType("MerkleProof", Sequence[Hash32])
-
-
-def get_root(tree: MerkleTree) -> Hash32:
-    """
-    Get the root hash of a Merkle tree.
-    """
-    return tree[0][0]
-
-
-def get_branch_indices(node_index: int, depth: int) -> Iterable[int]:
-    """
-    Get the indices of all ancestors up until the root for a node with a given depth.
-    """
-    return tuple(take(depth, iterate(lambda index: index // 2, node_index)))
+from .common import (  # noqa: F401
+    _calc_parent_hash,
+    _hash_layer,
+    get_branch_indices,
+    get_root,
+    MerkleTree,
+    MerkleProof,
+)
 
 
 def get_merkle_proof(tree: MerkleTree, item_index: int) -> Iterable[Hash32]:
@@ -69,13 +58,6 @@ def get_merkle_proof(tree: MerkleTree, item_index: int) -> Iterable[Hash32]:
     )
 
 
-def _calc_parent_hash(left_node: Hash32, right_node: Hash32) -> Hash32:
-    """
-    Calculate the parent hash of a node and its sibling.
-    """
-    return hash_eth2(left_node + right_node)
-
-
 def verify_merkle_proof(root: Hash32,
                         item: Union[bytes, bytearray],
                         item_index: int,
@@ -95,16 +77,6 @@ def verify_merkle_proof(root: Hash32,
         leaf,
     )
     return proof_root == root
-
-
-def _hash_layer(layer: Sequence[Hash32]) -> Iterable[Hash32]:
-    """
-    Calculate the layer on top of another one.
-    """
-    return tuple(
-        _calc_parent_hash(left, right)
-        for left, right in partition(2, layer)
-    )
 
 
 def calc_merkle_tree(items: Sequence[Union[bytes, bytearray]]) -> MerkleTree:
